WebSockets/tests/submissions/Opera/opening-handshake/001.html

<!doctype html>
<title>WebSockets: invalid handshake</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../constants.js></script>
<div id=log></div>
<script>
setTestTimeout(10000, 'got no close event');
async_test(function(t) {
  var ws = new WebSocket(SCHEME_AND_DOMAIN+'/ws/invalid');
  ws.onclose = function(e) {
    assert_equals(e.wasClean, false);
    
    ws.onclose = t.step_func(function() {assert_unreached()});
    
    setTimeout(t.step_func(function() {t.done();}), 50)
  }
   ws.onmessage = ws.onopen = t.step_func(function() {assert_unreached()});
});
</script>
